requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
Brlapi==0.6.7
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
distro==1.3.0
distro-info==0.21
entrypoints==0.3
gitdb==4.0.2
GitPython==3.1.0
httpie==0.9.8
httplib2==0.11.3
idna==2.6
isort==4.3.15
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
louis==3.8.0
Mako==1.0.7
MarkupSafe==1.1.0
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.2
olefile==0.46
pandas==1.0.3
patsy==0.5.1
pexpect==4.6.0
Pillow==5.4.1
psutil==5.5.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
pylint==2.3.1
pyparsing==2.4.6
PySimpleSOAP==1.16.2
pysmbc==1.0.15.6
python-apt==1.8.4.1
python-dateutil==2.8.1
python-debian==0.1.35
python-debianbts==2.8.2
python-pam==1.8.4
python-xapp==1.2.0
python-xlib==0.23
pytz==2019.3
pyxdg==0.25
reportbug===7.5.3-deb10u1
reportlab==3.5.13
requests==2.21.0
scipy==1.4.1
SecretStorage==2.3.1
setproctitle==1.1.10
six==1.12.0
smmap==3.0.1
statsmodels==0.11.1
typed-ast==1.3.1
unattended-upgrades==0.1
urllib3==1.24.1
virtualenv==15.1.0
wrapt==1.11.1
